subcommand status (#29)
* first sketch of status subcommand using system's ping

I chose to start with a simple sketch by having N threads each using the system "ping" command to check if machines are alive.
Using the systems ping command avoids dealing with sockets to implement ICMP directly, which runs into issues witrh Administrator/root privileges due to RAW sockets. Therefore the practical choice is to go with the lame ping command...

Moreover, once zig 0.16 is out, hopefully with the new async implementation, I'll rewrite this using asyncrony.

Knows bug: if using --live the results come in, obsiously, in a random order, since the threads themselves are writing the result out and are nto instead collected properly in the subcommand function... It's justa  sketch to check if it's doable :D

* disable live status (ping continously feature) for now

likely until I put async in this will stay off. results still come in un-ordered which is annoying but usable