Port to zig 0.15 (#27)
* fix build.zig deprecated .root_source_file

* partial port to 0.15.1 (array list init and printing)

serialize to ZON is broken with io.writer, gotta fix that

* fixed serializing zon to a file writer with the new interface

Also contains a commented example of how to write to stdout with the new std.io.writer interface which requires a buffer.

Also fixed the printing of net.Address with the new format flag {f} in strings.

Should move all "printf debugging" to a proper log at some point...

* Update main.yml

* replace deprecated std.time.sleep()

* Update README required Zig version for building